normalized and refactored
